Refactor user controller

+++ voicereader/user/controller.py
@@ -2,8 +2,13 @@ import json
 import ast
 import datetime
 import time
+
 from flask import Blueprint, request, jsonify
 from flask_pymongo import PyMongo
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from voicereader.constant import action_result
+from voicereader.constant import msg_json, MSG_NOT_EQUAL_IDENTITY, MSG_NOT_FOUND_ELEMENT
+
 from pymongo import TEXT
 from pymongo.errors import DuplicateKeyError
 from bson import ObjectId
@@ -21,21 +26,24 @@ def get_user_api(app):
 
 
 @_user_api.route('/<user_id>', methods=['GET'])
+@jwt_required
 def fetch_by_id(user_id):
     """ Function to fetch the users. """
+    if get_jwt_identity() != user_id:
+        return action_result.bad_request(msg_json(MSG_NOT_EQUAL_IDENTITY))
+
     try:
         # Fetch all the record(s)
         records_fetched = mongo.db.users.find_one({"_id": ObjectId(user_id)})
 
         if records_fetched is not None:
-            return jsonify(records_fetched), 200
+            return action_result.ok(jsonify(records_fetched))
         else:
-            return "", 404
+            return action_result.not_found(msg_json(MSG_NOT_FOUND_ELEMENT))
     except Exception as ex:
         # Error while trying to fetch the resource
         # Add message for debugging purpose
-        print(ex)
-        return "", 500
+        return action_result.internal_server_error(msg_json(str(ex)))
 
 
 @_user_api.route('', methods=['GET'])
@@ -43,18 +51,16 @@ def fetch_user_id_by_fcm_uid():
     fcm_uid = request.args.get('fcm_uid')
 
     if fcm_uid is None:
-        return '', 400
+        return action_result.bad_request(msg_json('not exists fcm uid'))
 
     try:
         records_fetched = mongo.db.users.find_one({"fcm_uid": fcm_uid})
+        if records_fetched is None:
+            return action_result.not_found(msg_json(MSG_NOT_FOUND_ELEMENT))
 
-        if records_fetched is not None:
-            return str(records_fetched['_id'])
-        else:
-            return '', 404
+        return action_result.ok(str(records_fetched['_id']))
     except Exception as ex:
-        print(ex)
-        return '', 500
+        return action_result.internal_server_error(msg_json(str(ex)))
 
 
 @_user_api.route('', methods=['POST'])
@@ -70,7 +76,7 @@ def add():
         except Exception as ex:
             # Bad request as request body is not available
             # Add message for debugging purpose
-            return ex, 400
+            return action_result.bad_request(msg_json(str(ex)))
 
         body['_id'] = ObjectId()
         body['email'] = decode_token['email']
@@ -79,23 +85,28 @@ def add():
 
         mongo.db.users.insert(body)
 
-        return jsonify(body), 201
+        return action_result.created(jsonify(body))
     except Error as ex:
-        return str(ex), 401
+        return action_result.unauthorized(msg_json(str(ex)))
     except ValueError as ex:
-        return str(ex), 401
+        return action_result.unauthorized(msg_json(str(ex)))
     except DuplicateKeyError as ex:
-        return str(ex), 409
+        return action_result.conflict(msg_json(str(ex)))
 
 
 @_user_api.route('/<user_id>/photo', methods=['POST'])
+@jwt_required
 def upload_photo(user_id):
     pass
 
 
 @_user_api.route('/<user_id>', methods=['PUT'])
+@jwt_required
 def update(user_id):
     """ Function to update the user. """
+    if get_jwt_identity() != user_id:
+        return action_result.bad_request(msg_json(MSG_NOT_EQUAL_IDENTITY))
+
     try:
         # Get the value which needs to be updated
         try:
@@ -103,8 +114,7 @@ def update(user_id):
         except Exception as ex:
             # Bad request as the request body is not available
             # Add message for debugging purpose
-            print(ex)
-            return "", 400
+            return action_result.bad_request(msg_json(str(ex)))
 
         # Updating the user
         records_updated = mongo.db.users.update_one({"_id": ObjectId(user_id)}, body)
@@ -112,21 +122,24 @@ def update(user_id):
         # Check if resource is updated
         if records_updated.modified_count > 0:
             # Prepare the response as resource is updated successfully
-            return "", 200
+            return action_result.ok(jsonify(body))
         else:
             # Bad request as the resource is not available to update
             # Add message for debugging purpose
-            return "", 404
+            return action_result.not_found(msg_json(MSG_NOT_FOUND_ELEMENT))
     except Exception as ex:
         # Error while trying to update the resource
         # Add message for debugging purpose
-        print(ex)
-        return "", 500
+        return action_result.internal_server_error(msg_json(str(ex)))
 
 
 @_user_api.route('/<user_id>', methods=["DELETE"])
+@jwt_required
 def remove(user_id):
     """ Function to remove the user. """
+    if get_jwt_identity() != user_id:
+        return action_result.bad_request(msg_json(MSG_NOT_EQUAL_IDENTITY))
+
     try:
         # Delete the user matched
         record_deleted = mongo.db.users.delete_one({"_id": ObjectId(user_id)})
@@ -135,11 +148,12 @@ def remove(user_id):
         if record_deleted.deleted_count > 0:
             # We return 204 No Content to imply resource updated successfully without returning
             # the deleted entity.
-            return "", 204
+            return action_result.no_content({
+                "user_id": user_id
+            })
         else:
             # Entity not found, perhaps already deleted, return 404
-            return "", 404
+            return action_result.not_found(msg_json(MSG_NOT_FOUND_ELEMENT))
     except Exception as ex:
         # Something went wrong server side, so return Internal Server Error.
-        print(ex)
-        return "", 500
+        return action_result.internal_server_error(msg_json(str(ex)))
